% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ecb_codelist}
\alias{ecb_codelist}
\title{Returns available code lists}
\usage{
ecb_codelist(agency = NULL, id = NULL)
}
\arguments{
\item{agency}{\code{character(1)} the agency to query. Defaut \code{NULL}.}

\item{id}{\code{character(1)} id to query. Default \code{NULL}.}
}
\value{
A data.frame with the available code lists. The columns are:
\item{id}{The id of the code list}
\item{name}{The name of the code list}
}
\description{
Returns available code lists
}
\examples{
ecb_codelist()
# or filter by id
ecb_codelist(id = "CLI_EONIA_BANK")
}
\references{
\url{https://data.ecb.europa.eu/help/api/metadata}
}
\seealso{
Other metadata: 
\code{\link{ecb_data_structure}()}
}
\concept{metadata}
