% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ecb_data_structure}
\alias{ecb_data_structure}
\title{Returns available data structures}
\usage{
ecb_data_structure(agency = NULL, id = NULL)
}
\arguments{
\item{agency}{\code{character(1)} the agency to query. Defaut \code{NULL}.}

\item{id}{\code{character(1)} id to query. Default \code{NULL}.}
}
\value{
A \code{data.frame()} with the available data structures.
The columns are:
\item{id}{The id of the data structure}
\item{name}{The name of the data structure}
}
\description{
Returns available data structures
}
\examples{
ecb_data_structure()
# or filter by id
ecb_data_structure(id = "ECB_BCS1")
}
\references{
\url{https://data.ecb.europa.eu/help/api/metadata}
}
\seealso{
Other metadata: 
\code{\link{ecb_codelist}()}
}
\concept{metadata}
