% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecbr-package.R
\docType{package}
\name{ecbr-package}
\alias{ecbr}
\alias{ecbr-package}
\title{ecbr: European Central Bank (ECB) API Wrapper}
\description{
Download data from the ECB API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://m-muecke.github.io/ecbr}
  \item \url{https://github.com/m-muecke/ecbr}
  \item Report bugs at \url{https://github.com/m-muecke/ecbr/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Muecke \email{muecke.maximilian@gmail.com} (\href{https://orcid.org/0009-0000-9432-9795}{ORCID})

}
\keyword{internal}
